requirements.txt
openai==1.55.1
Flask==3.1.0
PyMuPDF==1.24.14
python-dotenv==1.0.1
requests==2.32.3